+++ src/main/java/seedu/fitchasers/gym/Machine.java
@@ -1,4 +1,4 @@
-package seedu.fitchasers;
+package seedu.fitchasers.gym;
 
 import java.util.List;
