+++ src/main/java/seedu/fitchasers/user/WeightRecord.java
@@ -1,4 +1,4 @@
-package seedu.fitchasers;
+package seedu.fitchasers.user;
 
 import java.io.Serializable;
 import java.time.LocalDate;
