+++ src/test/java/seedu/fitchasers/UITest.java
@@ -11,6 +11,7 @@ import java.io.PrintStream;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import seedu.fitchasers.UI.UI;
 
 
 /**
