+++ src/test/java/seedu/fitchasers/WeightManagerTest.java
@@ -8,6 +8,9 @@ import java.util.List;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import seedu.fitchasers.user.Person;
+import seedu.fitchasers.user.WeightManager;
+import seedu.fitchasers.user.WeightRecord;
 
 class WeightManagerTest {
