+++ src/test/java/seedu/fitchasers/WorkoutTest.java
@@ -2,6 +2,9 @@ package seedu.fitchasers;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import seedu.fitchasers.workouts.Exercise;
+import seedu.fitchasers.workouts.Workout;
+
 import java.time.LocalDateTime;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
